display message if user is not admin, add some fixes and newlines to files

+++ src/modules/user/components/UsersComponent.js
@@ -1,12 +1,15 @@
-import React, {useEffect} from "react";
+import React, {useEffect, useState} from "react";
 import {useDispatch, useSelector} from "react-redux";
 import {deleteUserByAdminAction, getUsersListAction} from "../store/actions";
 import ViewUsersComponent from "./ViewUsersComponent";
-import {getProgramsListAction} from "../../program/store/actions";
+import useUser from "../hooks/useUser";
 
 const UsersComponent = () => {
     const dispatch = useDispatch();
     const { list } = useSelector((store) => store.domain.users.data);
+    const [displayError, setDisplayError] = useState(false);
+    const [loggedUserIsAdmin, setLoggedUserIsAdmin] = useState(null);
+    const isAdmin = useUser().isAdmin;
 
     const getUserList = () => {
         const paginationFilter = {
@@ -24,15 +27,24 @@ const UsersComponent = () => {
        };
 
        dispatch(getUsersListAction(paginationFilter));
-    }, [dispatch]);
+        setLoggedUserIsAdmin(isAdmin);
+    }, [dispatch, isAdmin]);
 
     const deleteUserByAdmin = (id) => {
-        dispatch(deleteUserByAdminAction(id)).then(() => getUserList());
+
+        if(loggedUserIsAdmin){
+            setDisplayError(false);
+            dispatch(deleteUserByAdminAction(id)).then(() => getUserList());
+        }else{
+            setDisplayError(true);
+            setTimeout(() => setDisplayError(false), 5000);
+        }
     }
 
     return (
         <>
             <h4>All users</h4>
+            { displayError ? <div className="container-fluid alert alert-danger">You must be admin to delete users</div> : null }
             <ViewUsersComponent {...list} deleteUserByAdmin={deleteUserByAdmin} />
         </>
     );
